fix: musl libc

+++ cross-sysroot-linux-generator/organize_sysroot.sh
@@ -31,14 +31,18 @@ rm -f ${SYSROOT}/lib/libitm.spec
 rm -f ${SYSROOT}/usr/lib/libBrokenLocale.a
 rm -f ${SYSROOT}/usr/lib/libanl.a
 rm -f ${SYSROOT}/usr/lib/libc.a
-rm -f ${SYSROOT}/usr/lib/libcrypt.a
-rm -f ${SYSROOT}/usr/lib/libdl.a
-rm -f ${SYSROOT}/usr/lib/libm.a
 rm -f ${SYSROOT}/usr/lib/libnsl.a
-rm -f ${SYSROOT}/usr/lib/libpthread.a
-rm -f ${SYSROOT}/usr/lib/libresolv.a
-rm -f ${SYSROOT}/usr/lib/librt.a
-rm -f ${SYSROOT}/usr/lib/libutil.a
+
+if [[ "${SYSROOT}" =~ ^.*gnu.*$ ]]; then
+  rm -f ${SYSROOT}/usr/lib/libcrypt.a
+  rm -f ${SYSROOT}/usr/lib/libdl.a
+  rm -f ${SYSROOT}/usr/lib/libm.a
+  rm -f ${SYSROOT}/usr/lib/libpthread.a
+  rm -f ${SYSROOT}/usr/lib/libresolv.a
+  rm -f ${SYSROOT}/usr/lib/librt.a
+  rm -f ${SYSROOT}/usr/lib/libutil.a
+fi
+
 
 mv tmp/${CROSS_TRIPLE}/include ${SYSROOT}/
 cd ${SYSROOT}/include; CXXDIR="$(dirname c++/*/${CROSS_TRIPLE})"; cd -;
